Replace 'r#type' with 'pattern_type'.
The 'r#' syntax, which allows keywords to be used as identifiers, is an
esoteric feature of the language. This change will make the code more
accessible to those not aware of the feature and will avoid tripping up
developer tools that aren't aware of it (e.g. coc-rls's identifier rename).